+++ netpyne/analysis/__init__.py
@@ -22,13 +22,14 @@ from .wrapper import plotData
 # Import utils methods
 # -------------------------------------------------------------------------------------------------------------------
 from .utils import exception, _showFigure, _saveFigData, getCellsInclude, getCellsIncludeTags, _roundFigures, \
-     _smooth1d, syncMeasure, invertDictMapping
+     _smooth1d, syncMeasure, invertDictMapping, checkAvailablePlots
 
 
 # -------------------------------------------------------------------------------------------------------------------
 # Import connectivity-related functions
 # -------------------------------------------------------------------------------------------------------------------
-from .network import plotConn, _plotConnCalculateFromSim, _plotConnCalculateFromFile, plot2Dnet, plotShape, calculateDisynaptic
+from .network import plotConn, _plotConnCalculateFromSim, _plotConnCalculateFromFile, plot2Dnet, plotShape, calculateDisynaptic, \
+    plot2Dfiring
 
 
 # -------------------------------------------------------------------------------------------------------------------